requirements.txt
python-dotenv==1.0.1
psycopg2-binary==2.9.10
djangorestframework==3.15.2
djangorestframework-simplejwt==5.5.0